backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-multipart
pymupdf
Pillow
numpy
pandas
matplotlib
seaborn
scipy
fpdf2
